agent_bar_v2/subagents/retail/intelligent_inventory_manager/pyproject.toml
[project]
name = "intelligent-inventory-manager"
version = "0.0.1"
description = "An agent that can search for retail products using a BQ database"
authors = [{ name = "Theo Haddad", email = "theohaddad@google.com" }]
classifiers = [
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.13",
]
keywords = ['agent-catalog', 'adk', 'agent', 'retail', 'product', 'search']
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.3.0",
    "google-cloud-bigquery>=3.35.1",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.34.3",
]

[build-system]
requires = [
  "setuptools >= 40.9.0",
]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "ipython>=8.36.0",
    "pytest>=8.3.5",
    "pre-commit>=4.2.0",
    "ruff>=0.11.5",
    "deptry>=0.23.0",
    "pytest-cov>=4.0.0",
    "python-dotenv>=1.1.1",
    "pytest-asyncio>=1.0.0",
    "basedpyright>=1.31.3",
]

[tool.agent_packs]
template = "adk"

[tool.basedpyright]
reportAny = false
reportExplicitAny = false
reportMissingTypeStubs = false
reportUnknownMemberType = false
reportUnknownVariableType = false
typeCheckingMode = "basic"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow",
    "infra_ready: marks tests as requiring cloud infra to be ready",
]

[tool.ruff]
line-length = 120
fix = true

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["src"]
omit = ["server.py"]

[tool.deptry.per_rule_ignores]
DEP002 = ["uvicorn"]